Made the app generate the fractal asynchronously | switched to x64 sfml
Switched the stepping of the pixel from depending on the framerate to being independent of it. This means that the fractal generates a lot faster, but the program now uses a ton of memory (will get fixed soon)

Large canvas sizes (over 8192 px) would cause the file saving to crash due to the massive amount of memory used (upwards of 2.1 GB on 10k x 10k canvas size). This was "fixed" by changing the program to x64. A real fix would be optimizing the actual data structures - which I will be getting to soon